+++ spec/app/models/doctor.rb
@@ -1,0 +1,9 @@
+class Doctor < Person
+  field :specialty
+  has_and_belongs_to_many :users, :validate => false, :inverse_of => nil
+
+  def specialty=(text)
+    users.push(User.new)
+    super
+  end
+end