+++ spec/unit/mongoid/document_spec.rb
@@ -331,6 +331,21 @@ describe Mongoid::Document do
       person.title.should == "Sir"
     end
 
+    context "when accessing a relation from an overridden setter" do
+
+      let(:doctor) do
+        Doctor.new(:specialty => "surgery")
+      end
+
+      it "allows access to the relation" do
+        doctor.users.first.should be_a(User)
+      end
+
+      it "properly allows super calls" do
+        doctor.specialty.should eq("surgery")
+      end
+    end
+
     context "when initialize callbacks are defined" do
 
       context "when accessing attributes" do