+++ spec/unit/mongoid/fields/internal/foreign_keys/object_spec.rb
@@ -2,6 +2,23 @@ require "spec_helper"
 
 describe Mongoid::Fields::Internal::ForeignKeys::Object do
 
+  describe "#foreign_key?" do
+
+    let(:field) do
+      described_class.instantiate(
+        :vals,
+        :metadata => Person.relations["posts"],
+        :type => Object,
+        :default => [],
+        :identity => true
+      )
+    end
+
+    it "returns true" do
+      field.should be_foreign_key
+    end
+  end
+
   describe "#serialize" do
 
     context "when the array is object ids" do